cilium-health: Refactor output printing to pkg/
Shift the formatting of results to pkg/health/client/ so that a
subsequent patch can reuse the formatting code.
